requirements.txt
python-dotenv==1.1.0
<<<<<<< HEAD
<<<<<<< HEAD
requests==2.32.3
python-telegram-bot==13.15
=======
requests==2.32.3
=======
requests==2.32.3
python-telegram-bot==13.15

